feat: rfc styling with scoped upstream styles (mostly)
