+++ client/assets/css/rfc-plaintext.css
@@ -1,0 +1,28 @@
+/**
+
+This is derived from 
+
+    https://github.com/ietf-tools/datatracker/blob/2bf633bf70c40b9cb6baf428901615a0403e1ea5/ietf/static/css/document_html.scss#L175
+
+*/
+/* A lot of plaintext documents seem to have line lengths >72ch.
+// To handle that, we calculate with 80ch here and adjust some of the
+// font sizes down accordingly. */
+pre {
+    width: 80ch;
+    white-space: pre-wrap;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+    white-space: pre;
+    display: inline;
+}
+
+.newpage {
+    margin-top: -1.25em;
+}
