+++ client/package.json
@@ -52,6 +52,7 @@
     "nitropack": "2.10.4",
     "nuxt": "3.15.4",
     "pinia-plugin-persistedstate": "4.3.0",
+    "postcss-nested-import": "1.3.0",
     "reka-ui": "2.2.1",
     "remark-heading-id": "1.0.1",
     "sharp": "0.34.2",
