+++ client/utilities/rfc.ts
@@ -212,4 +212,5 @@ export type RfcBucketHtmlDocument = {
   rfc: RfcCommon
   tableOfContents?: RfcEditorToc
   documentHtml: string
+  documentHtmlType: 'plaintext' | 'xml2rfc'
 }
